refactor(dbconnection): put connection logic into db folder
